Maze Collision Start
Started work on Maze collisions. As of right now, the app “crashes”
when a collision occurs. Have tried everything I can to figure out why
this is, but still don’t know. Feel free to test it out and see if you
can figure out why it’s crashing. I’ll continue to work on it.